# mydoc/export/agent-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: VCP Agent Gateway API
  version: 1.0.0
  description: |-
    VCP Agent Gateway 的正式 Native published contract。

    说明：
    - 当前对外协议版本为 `v1`，发布阶段为 `ga`。
    - 该文档覆盖当前正式开放的 `health / capabilities / agents / metrics / memory / context / tools / jobs / events` 资源。
    - 推荐优先使用独立 Gateway 凭证接入；Basic Auth 与 `admin_auth` Cookie 仅作为过渡兼容链路记录。
    - 该契约与 `routes/agentGatewayRoutes.js`、M7-M11 主 specs 以及 contract validation 保持对齐。
  contact:
    name: VCP Team
  x-gateway-version: v1
  x-release-stage: ga
  x-vcp-version: 1.0.0
servers:
  - url: http://localhost:3000
    description: 本地默认端口
  - url: http://localhost:8080
    description: 备用端口
...
security:
  - gatewayKeyHeader: []
  - gatewayBearerAuth: []
  - basicAuth: []
  - adminAuthCookie: []
paths:
...
  /agent_gateway/agents:
    get:
      tags:
        - Agents
      summary: 获取当前 registry 中的 agent 列表
      operationId: listAgentGatewayAgents
      parameters:
        - $ref: "#/components/parameters/RequestIdQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/RuntimeQuery"
      responses:
        "200":
          $ref: "#/components/responses/AgentListSuccess"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: 获取单个 agent 的详细定义
      operationId: getAgentGatewayAgentDetail
      parameters:
        - $ref: "#/components/parameters/AgentIdPath"
        - $ref: "#/components/parameters/RequestIdQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/RuntimeQuery"
      responses:
        "200":
          $ref: "#/components/responses/AgentDetailSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/agents/{agentId}/render:
    post:
      tags:
        - Agents
      summary: 将单个 agent 渲染为可消费 prompt
      operationId: renderAgentGatewayAgent
      parameters:
        - $ref: "#/components/parameters/AgentIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRenderRequest"
      responses:
        "200":
          $ref: "#/components/responses/AgentRenderSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    gatewayKeyHeader:
      type: apiKey
      in: header
      name: x-agent-gateway-key
      description: 推荐的 Native Gateway 专用凭证。
    gatewayBearerAuth:
      type: http
      scheme: bearer
      description: 与 gateway key 等价的 Bearer 认证输入。
    basicAuth:
      type: http
      scheme: basic
      description: 过渡期兼容的管理后台 Basic Auth。
    adminAuthCookie:
      type: apiKey
      in: cookie
      name: admin_auth
      description: 过渡期兼容的管理态 Cookie。
  headers:
    XRequestId:
      description: 请求 ID
      schema:
        type: string
    XGatewayVersion:
      description: Agent Gateway 运行时版本
      schema:
        type: string
        example: v1
    XTraceId:
      description: Agent Gateway 操作级 trace 标识
      schema:
        type: string
        example: agwop_1234567890abcdef123456
  parameters:
    AgentIdPath:
      name: agentId
      in: path
      required: true
      schema:
        type: string
...
    RequestIdQuery:
      name: requestId
      in: query
      required: false
      schema:
        type: string
...
    SourceQuery:
      name: source
      in: query
      required: false
      schema:
        type: string
    RuntimeQuery:
      name: runtime
      in: query
      required: false
      schema:
        type: string
  responses:
...
    AgentListSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentListEnvelope"
    AgentDetailSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentDetailEnvelope"
    AgentRenderSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentRenderEnvelope"
          examples:
            renderedPrompt:
              summary: final rendered prompt output
              value:
                success: true
                data:
                  agentId: Ariadne
                  alias: Ariadne
                  sourceFile: Agent/Ariadne.txt
                  renderedPrompt: |-
                    你是阿里阿德涅，当前项目是 VCPToolBox。
                    记忆片段：上周完成了 gateway render contract 收口。
                  dependencies:
                    agents: []
                    toolboxes: []
                    variables:
                      - VarSystemInfo
                    ragBlocks:
                      - "[[阿里阿德涅日记本::Time::TagMemo]]"
                    metaThinkingBlocks: []
                    asyncResults: []
                  unresolved: []
                  warnings: []
                  truncated: false
                  renderMeta:
                    memoryRecallApplied: true
                    recallSources:
                      - tagmemo
                    truncated: false
                    filteredByPolicy: false
                    unresolvedCount: 0
                    variableKeys:
                      - VarSystemInfo
                  meta:
                    model: gpt-4.1
                    rawSize: 2400
                    renderedSize: 2800
                    variableKeys:
                      - VarSystemInfo
                meta:
                  requestId: req-agent-render-001
                  durationMs: 12
                  gatewayVersion: v1
                  traceId: agwop_renderexample1234567890
                  operationName: agents.render
...
    NotFound:
      description: 目标 agent、tool、job 或资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    GatewayMeta:
      type: object
      required:
        - requestId
        - durationMs
        - gatewayVersion
      properties:
        requestId:
          type: string
        durationMs:
          type: integer
        gatewayVersion:
          type: string
          example: v1
        authMode:
          type: string
          example: gateway_key
        authSource:
          type: string
          example: x-agent-gateway-key
        gatewayId:
          type: string
          example: vcp-gateway
        traceId:
          type: string
          example: agwop_1234567890abcdef123456
        operationName:
          type: string
          example: tool.invoke
        retryAfterMs:
          type: integer
        toolStatus:
          type: string
          enum:
            - completed
            - accepted
            - waiting_approval
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - code
        - meta
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
        code:
          type: string
          example: AGW_FORBIDDEN
        details:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    RequestContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
          example: native
    AuthContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
        adapter:
          type: string
        gatewayId:
          type: string
        authMode:
          type: string
        authSource:
          type: string
        roles:
          type: array
          items:
            type: string
    Message:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
    MemoryTarget:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        type:
          type: string
          example: diary
        allowed:
          type: boolean
    ToolDescriptor:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        pluginType:
          type: string
        distributed:
          type: boolean
        approvalRequired:
          type: boolean
        timeoutMs:
          type: integer
        description:
          type: string
        inputSchema:
          type: object
          additionalProperties: true
        invocationCommands:
          type: array
          items:
            type: object
            additionalProperties: true
...
    DefaultPolicies:
      type: object
      properties:
        toolNames:
          type: array
          items:
            type: string
        memoryTargetIds:
          type: array
          items:
            type: string
    AgentPolicyHints:
      type: object
      properties:
        toolNames:
          type: array
          items:
            type: string
        memoryTargetIds:
          type: array
          items:
            type: string
        contextSupported:
          type: boolean
        memoryWriteSupported:
          type: boolean
        jobsSupported:
          type: boolean
        eventsSupported:
          type: boolean
    AgentListItem:
      type: object
      properties:
        agentId:
          type: string
        alias:
          type: string
        sourceFile:
          type: string
        exists:
          type: boolean
        mtime:
          type: string
          format: date-time
          nullable: true
        hash:
          type: string
        summary:
          type: string
        defaultPolicies:
          $ref: "#/components/schemas/DefaultPolicies"
        capabilityHints:
          $ref: "#/components/schemas/AgentPolicyHints"
    AgentListData:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentListItem"
    AgentListEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AgentListData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    PlaceholderSummary:
      type: object
      properties:
        total:
          type: integer
        agents:
          type: array
          items:
            type: string
        toolboxes:
          type: array
          items:
            type: string
        variables:
          type: array
          items:
            type: string
        ragBlocks:
          type: integer
        metaThinkingBlocks:
          type: integer
        asyncResults:
          type: integer
    PromptDependencies:
      type: object
      properties:
        agents:
          type: array
          items:
            type: string
        toolboxes:
          type: array
          items:
            type: string
        variables:
          type: array
          items:
            type: string
        ragBlocks:
          type: array
          items:
            type: string
        metaThinkingBlocks:
          type: array
          items:
            type: string
        asyncResults:
          type: array
          items:
            type: object
            properties:
              pluginName:
                type: string
              requestId:
                type: string
    AgentDetail:
      type: object
      properties:
        agentId:
          type: string
        alias:
          type: string
        sourceFile:
          type: string
        exists:
          type: boolean
        mtime:
          type: string
          format: date-time
          nullable: true
        hash:
          type: string
        summary:
          type: string
        defaultPolicies:
          $ref: "#/components/schemas/DefaultPolicies"
        capabilityHints:
          $ref: "#/components/schemas/AgentPolicyHints"
        prompt:
          type: object
          properties:
            raw:
              type: string
            size:
              type: integer
            placeholderSummary:
              $ref: "#/components/schemas/PlaceholderSummary"
            dependencies:
              $ref: "#/components/schemas/PromptDependencies"
        accessibleTools:
          type: array
          items:
            $ref: "#/components/schemas/ToolDescriptor"
        accessibleMemoryTargets:
          type: array
          items:
            $ref: "#/components/schemas/MemoryTarget"
    AgentDetailEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AgentDetail"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    AgentRenderRequest:
      type: object
      properties:
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
        variables:
          type: object
          additionalProperties:
            type: string
        model:
          type: string
        maxLength:
          type: integer
        context:
          type: object
          additionalProperties: true
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    AgentRenderResult:
      type: object
      properties:
        agentId:
          type: string
        alias:
          type: string
        sourceFile:
          type: string
        renderedPrompt:
          type: string
        dependencies:
          $ref: "#/components/schemas/PromptDependencies"
        unresolved:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        truncated:
          type: boolean
        renderMeta:
          $ref: "#/components/schemas/AgentRenderMeta"
        meta:
          type: object
          additionalProperties: true
    AgentRenderMeta:
      type: object
      properties:
        memoryRecallApplied:
          type: boolean
        recallSources:
          type: array
          items:
            type: string
        truncated:
          type: boolean
        filteredByPolicy:
          type: boolean
        unresolvedCount:
          type: integer
        variableKeys:
          type: array
          items:
            type: string
    AgentRenderEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AgentRenderResult"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
...